cat-vision-ia/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.6.1
httpx==0.28.1
pandas==2.2.3
pypdf==5.1.0
PyMuPDF==1.24.14
pdfplumber==0.11.4
pillow==11.0.0
pytesseract==0.3.13
pdf2image==1.17.0
python-dateutil==2.9.0.post0
unidecode==1.3.8
rapidfuzz==3.10.1
pytest==8.3.4
pytest-asyncio==0.24.0

# Opcional: SmolDocling via Docling (pesado; instale só se necessário)
# docling>=2.0.0
